Alterações em endpoints de deletar prioridade

+++ src/main/java/com/boardhub/BoardPi/services/PrioridadeService.java
@@ -3,7 +3,9 @@ package com.boardhub.BoardPi.services;
 import com.boardhub.BoardPi.entities.Prioridade;
 import com.boardhub.BoardPi.entities.Projeto;
 import com.boardhub.BoardPi.repositories.PrioridadeRepository;
+import com.boardhub.BoardPi.repositories.TarefaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,6 +15,9 @@ public class PrioridadeService {
 
     @Autowired
     private PrioridadeRepository prioridadeRepository;
+    @Autowired
+    private TarefaRepository tarefaRepository;
+
     public List<Prioridade> getPrioridades() {
         return prioridadeRepository.findAll();
     }
@@ -26,7 +31,12 @@ public class PrioridadeService {
         return prioridadeRepository.save(prioridade);
     }
     public void deletePrioridade(long id) {
-        prioridadeRepository.deleteById(id);
+        Prioridade p = getPrioridade(id);
+        if(p != null) {
+            tarefaRepository.setPrioridadeNull(p);
+            prioridadeRepository.deleteById(id);
+        }
+
     }
     public List<Prioridade> getPorProjeto(Projeto projeto) {
         return prioridadeRepository.findByProjeto(projeto);
